requirements.txt
langchain
streamlit
openai
tiktoken
python-dotenv
pinecone-client
scipy
plotly
umap-learn==0.5.3
scikit-learn
setuptools